applications/consumer/requirements.txt
# Kafka
kafka-python-ng==2.2.2

# PostgreSQL
psycopg2-binary==2.9.9

# Prometheus metrics
prometheus-client==0.20.0

# Logging
python-json-logger==2.0.7

# Utilities
pydantic==2.6.4
pydantic-settings==2.2.1

# OpenTelemetry - Tracing
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation==0.45b0
